Tile locations idea implemented
Tile objects now contain pointers to the tiles above, below, to the left and to the right. New addTile function takes a reference tile, the tile to be added and asks the user where to place the tile relative to the reference tile. Also board constructer now uses randomTile function and new starterTiles function