added a bag lol
hey guys did u know that the tokens are all in a bag with 100 of them in it, i did not. anyways its added. if wildlife tokens are culled or not placed, they go back in the back of the bag. also you apparently cannot have a 'hand' when placing a wildlife token, instead if u dont or can place it, it gets shuffled back into the back of the bag